pyproject.toml
[project]
name = "10-making-my-mom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "discord-py>=2.6.4",
    "faiss-cpu>=1.13.2",
    "google-genai>=1.61.0",
    "hydra-core>=1.3.2",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "openai>=2.16.0",
    "pydantic>=2.12.5",
    "sentence-transformers>=5.2.2",
]
